Expanded base/derived class hierarchy
