+++ MikolovJoulinChopraEtAl2015/python/base_rnn_graph3.py
@@ -1,4 +1,4 @@
-# Structurally Contrained Recurrent Network (SCRN) Model
+# Structurally Constrained Recurrent Network (SCRN) Model
 #
 # This gives an implementation of the SCRN model given in Mikolov et al. 2015, arXiv:1412.7753 [cs.NE], 
 # https://arxiv.org/abs/1412.7753 using Python and Tensorflow.
@@ -18,11 +18,12 @@ import numpy as np
 import tensorflow as tf
 
 # Local imports
+from base_rnn_graph import base_rnn_graph
 from batch_generator import batch_generator
 from log_prob import log_prob
 
-# Define base RNN TensorFlow graph class
-class base_rnn_graph3(object):
+# Define derived RNN TensorFlow graph class with saved hidden and state vectors
+class base_rnn_graph3(base_rnn_graph):
     
     # Graph constructor
     def __init__(self, num_gpus, hidden_size, state_size, vocabulary_size, num_training_unfoldings,
@@ -139,14 +140,6 @@ class base_rnn_graph3(object):
             # Validation prediction, replace with hierarchical softmax in the future
             self._validation_prediction = tf.nn.softmax(logits)
             
-    # Placeholder function for cell definition
-    def _cell(self):
-        print('Cell not defined')
-    
-    # Placeholder function to set up cell parameters
-    def _setup_cell_parameters(self):
-        print('Cell parameters not defined')
-    
     # Implements a tower to run part of a batch of training data on a GPU
     def _training_tower(self, i, tower, gpu):
         
@@ -191,108 +184,3 @@ class base_rnn_graph3(object):
             with tf.control_dependencies([self._validation_hidden_saved[tower].assign(hidden), 
                                           self._validation_state_saved[tower].assign(state)]):
                 return outputs
-            
-    # Train model parameters
-    def train(self, learning_rate, learning_decay, num_epochs, summary_frequency, training_text, validation_text):
-
-        # Generate training batches
-        print('Training Batch Generator:')
-        training_batches = []
-        for tower in range(self._num_towers):
-            training_batches.append(batch_generator(tower, training_text[tower], self._batch_size,
-                                                    self._num_training_unfoldings, self._vocabulary_size))
-        
-        # Generate validation batches
-        print('Validation Batch Generator:')
-        validation_batches = []
-        tower = 0
-        for tower in range(self._num_towers):
-            validation_batches.append(batch_generator(tower, validation_text[tower], 1,
-                                                      self._num_validation_unfoldings, self._vocabulary_size))
-        
-        # Training loop
-        batch_ctr = 0
-        epoch_ctr = 0
-        training_feed_dict = dict()
-        validation_feed_dict = dict()
-        with tf.Session(graph=self._graph, config=tf.ConfigProto(log_device_placement=True)) as session:
-        
-            session.run(self._initialization)
-            print('Initialized')
-
-            # Iterate over fixed number of training epochs
-            for epoch in range(num_epochs):
-
-                # Display the learning rate for this epoch
-                print('Epoch: %d  Learning Rate: %.2f' % (epoch+1, learning_rate))
-
-                # Training Step:
-
-                # Iterate over training batches
-                for tower in range(self._num_towers):
-                    training_batches[tower].reset_token_idx()
-                session.run(self._reset_training_state)
-                for batch in range(training_batches[0].num_batches()):
-
-                    # Get next training batch
-                    training_batches_next = []
-                    tower = 0
-                    for tower in range(self._num_towers):
-                        training_batches_next.append([])
-                        training_batches_next[tower] = training_batches[tower].next()
-                    batch_ctr += 1
-
-                    # Optimization
-                    training_feed_dict[self._learning_rate] = learning_rate
-                    for tower in range(self._num_towers):
-                        for i in range(self._num_training_unfoldings + 1):
-                            training_feed_dict[self._training_data[tower][i]] = training_batches_next[tower][i]
-                    session.run(self._optimize, feed_dict=training_feed_dict)
-
-                    # Summarize current performance
-                    if (batch+1) % summary_frequency == 0:
-                        cst = session.run(self._cost, feed_dict=training_feed_dict)
-                        print('     Total Batches: %d  Current Batch: %d  Cost: %.2f' % 
-                              (batch_ctr, batch+1, cst))
-                      
-                # Validation Step:
-        
-                # Iterate over validation batches
-                for tower in range(self._num_towers):
-                    validation_batches[tower].reset_token_idx()
-                session.run(self._reset_validation_state)
-                validation_log_prob_sum = 0
-                for _ in range(validation_batches[0].num_batches()):
-                    
-                    # Get next validation batch
-                    validation_batches_next = []
-                    tower = 0
-                    for tower in range(self._num_towers):
-                        validation_batches_next.append([])
-                        validation_batches_next[tower] = validation_batches[tower].next()
-                    
-                    # Validation
-                    validation_batches_next_label = []
-                    for tower in range(self._num_towers):
-                        validation_batches_next_label_tmp = []
-                        for i in range(self._num_validation_unfoldings):
-                            validation_feed_dict[self._validation_input[tower][i]] = validation_batches_next[tower][i]
-                            validation_batches_next_label_tmp.append(validation_batches_next[tower][i+1])
-                        validation_batches_next_label.append(validation_batches_next_label_tmp)
-                    validation_prediction = session.run(self._validation_prediction, feed_dict=validation_feed_dict)
-                    
-                    # Summarize current performance
-                    for tower in range(self._num_towers):
-                        for i in range(self._num_validation_unfoldings):
-                            validation_log_prob_sum = validation_log_prob_sum + \
-                                log_prob(validation_prediction[tower][i], validation_batches_next_label[tower][i])
-                    
-                # Calculation validation perplexity
-                N = self._num_towers*self._num_validation_unfoldings*validation_batches[0].num_batches()
-                perplexity = float(2 ** (-validation_log_prob_sum / N))
-                print('Epoch: %d  Validation Set Perplexity: %.2f' % (epoch+1, perplexity))
-
-                # Update learning rate
-                if epoch > 0 and perplexity > perplexity_last_epoch:
-                    learning_rate *= learning_decay
-                perplexity_last_epoch = perplexity
